+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * New cop `UnreachableCode` tracks unreachable code segments.
 * New cop `MethodInvocationParentheses` tracks unwanted braces in method invocations.
 * New cop `UnusedLocalVariable` tracks unused local variables for each scope.
+* New cop `ShadowingOuterLocalVariable` tracks use of the same name as outer local variables for block arguments or block local variables.
 * Add support for auto-correction of some offences with `-a`/`--auto-correct`.
 
 ### Changes